backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
python-dotenv
python-multipart
google-cloud-storage>=2.16.0